Changed tests to use HL7 CCD.xml because Kinsights was missing Encounters
HL7 xml has lab effectiveTimes which are near Daylight Saving time. Chrome/FF/Node (incorrectly) say a Date of 03/23/2000 is NOT in Daylight Saving time. Safari/PhantomJS say it is. Because our browser tests run in PhantomJS and our Node tests run in Node I had to create two different expected_output jsons. One for the browser and one for Node. Otherwise the tests would always fail in one context or the other because they disagree on the look of the timestamps.

In the long run this is a really poor way to test but it's the best we have at the moment. Ultimately we should be testing the individual BB methods and primarily the stuff in core.js. So, for the time being, we have two expected outputs :\