This commit gets all tests passing again except for the lab tests and the tests of the complete JSON (which fail because of the labs).
This involved:
1. Adding some fields that were correctly being parsed into the expected
   output for allergies and demographics
2. Changing the case for a couple of fields, which get normalized by
   codes.js now (instead of, presumably, following the case in the
   original XML)
2. Uncommenting several fields in the vitals parser, which appear to
   correctly parse but were commented out.

I also removed the split between expected_browser and expected_node
output, opting instead to just selectively update the dates that node
and phantomjs disagree about at the top of ccda_spec.js. I hope this
will make it easier to keep the tests up-to-date in the future.