Correcting the identifying ID for CCDAs
Turns out the ID we were using ("2.16.840.1.113883.10.20.22.1.2") just
describes the ContinuityOfCareDocument flavor of C-CDAs... but there are
several other valid types, such as ConsultationNote and
DischargeSummary. Prior to this change, BB.js would refuse to parse
those.

Based on my review of the spec, I believe "2.16.840.1.113883.10.20.22.1.1"
is the correct ID to check for. It's possible this is too broad and
maybe applies to other kinds of CDAs (specs are hard...) but it's
probably better to be generous than conservative here.