Recognize and split out declined immunizations from regular immunizations.
negationInd is a super important flag for immunizations, because if it
is true, the whole record is documenting that this vaccine WASN'T
administered... We were ignoring it. To avoid subjecting library users
to that same issue, we will output all declined immunizations in a
separate list called 'immunization_declines', so it will be hard to make
the same slip-up.

negationInd can exist on problems and some other sections, so we may
have to pay attention in other places. But immunizations seems like the
most likely/important usage.