Expand the parseDate routine in bb.js to actually attempt to parse the time info, rather than simply discarding it (#123).
This change started by including moment.js, using it to update the new
expected values, and then extending parseDate until it could match the
output of moment.js, at least for the sample files in our unit tests.
It also adds unit tests to parseDate, so that we can add specific test
cases as we find new deficincies.