Important usage clarification about examples
I think this should be added because the several of the example CCDA documents provided are rejected as not XML, even though the documents are considered valid by online XML validators. This led to some confusion, but I figured out that the parser code requires the first five characters of the CCDA document to be <?xml. I hope this helps others use the tools.